fixing cart behaviour

+++ src/app/components/cart/cart.component.ts
@@ -1,7 +1,7 @@
 // src/app/components/cart/cart.component.ts
 import { Component, OnInit } from '@angular/core';
 import { CartService } from '../../services/cart.service';
-import {  CartItem} from "../../entity/cart";
+import { CartItem } from "../../entity/cart";
 import { CheckoutService } from '../../services/checkout.service';
 import { ActivatedRoute, Router } from '@angular/router';
 @Component({
@@ -15,7 +15,7 @@ export class CartComponent implements OnInit {
 
   secretKey !: string;
 
-  constructor(private cartService: CartService, private checkoutService : CheckoutService, private router : Router,private route: ActivatedRoute,) {}
+  constructor(private cartService: CartService, private checkoutService: CheckoutService, private router: Router, private route: ActivatedRoute,) { }
 
   ngOnInit() {
     this.cartService.cart$.subscribe(items => {
@@ -27,14 +27,28 @@ export class CartComponent implements OnInit {
   }
 
   updateQuantity(item: CartItem, change: number) {
-    const newQuantity = item.quantity + change;
+    let newQuantity = item.quantity + change;
     if (newQuantity > 0 && newQuantity <= item.product.availableQuantity) {
-      this.cartService.updateQuantity(item.product.id, newQuantity);
+   
+      this.cartService.updateOrder({
+        product_id: item.product.id,
+        quantity: -change
+      }).subscribe(response => {
+        if (response) {
+          this.cartService.updateQuantity(item.product.id, newQuantity);
+        }
+      })
     }
   }
-
-  removeItem(productId: number) {
-    this.cartService.removeFromCart(productId);
+  removeItem(productId: number, quantity: number) {
+    this.cartService.updateOrder({
+      product_id: productId,
+      quantity: quantity
+    }).subscribe(response => {
+      if (response) {
+        this.cartService.removeFromCart(productId);
+      }
+    })
   }
 
   updateCart() {
@@ -43,14 +57,14 @@ export class CartComponent implements OnInit {
     console.log('Cart updated');
   }
 
-  createPaymentIntent(){
-    this.checkoutService.createPaymentIntent({amount : this.total * 10 , currency:'USD', receiptEmail:'mouad10cherrat@gmail.com'}).subscribe(
+  createPaymentIntent() {
+    this.checkoutService.createPaymentIntent({ amount: this.total * 10, currency: 'USD', receiptEmail: 'mouad10cherrat@gmail.com' }).subscribe(
       response => {
         this.secretKey = JSON.parse(response?.paymentIntent).client_secret;
         console.log('secret key +++ ' + this.secretKey);
-        this.router.navigate(['checkout'], {relativeTo: this.route , state: { secretKey: this.secretKey, amount: this.total }} );
+        this.router.navigate(['checkout'], { relativeTo: this.route, state: { secretKey: this.secretKey, amount: this.total } });
       });
 
   }
-  
+
 }
